Fix checkstyle errors and remove remaining tests out of BigQueryTableInserterTest.
