+++ sdks/java/io/google-cloud-platform/src/test/java/org/apache/beam/sdk/io/gcp/bigquery/BigQueryUtilTest.java
@@ -31,7 +31,6 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 
-import autovalue.shaded.com.google.common.common.collect.Lists;
 import com.google.api.services.bigquery.Bigquery;
 import com.google.api.services.bigquery.model.Table;
 import com.google.api.services.bigquery.model.TableCell;
@@ -51,8 +50,6 @@ import java.util.List;
 import org.apache.beam.sdk.io.gcp.bigquery.BigQueryServicesImpl.DatasetServiceImpl;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
-import org.apache.beam.sdk.transforms.SerializableFunction;
-import org.apache.beam.sdk.transforms.SimpleFunction;
 import org.hamcrest.Matchers;
 import org.junit.After;
 import org.junit.Assert;
@@ -446,7 +443,7 @@ public class BigQueryUtilTest {
 
     long totalBytes = 0;
     try {
-      List<TableRow> deadLetter = Lists.newArrayList();
+      List<TableRow> deadLetter = new ArrayList<>();
       totalBytes = datasetService.insertAll(ref, rows, ids);
       assertTrue(deadLetter.isEmpty());
     } finally {
